requirements.txt
altgraph==0.17.2
auto-py-to-exe==2.10.1
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
docutils==0.18
Eel==0.12.4
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.2
idna==3.3
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
pefile==2021.9.3
pycparser==2.20
Pygments==2.10.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pypiwin32==223
python-decouple==3.5
pywin32==302
pywin32-ctypes==0.2.0
requests==2.26.0
urllib3==1.26.7
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
